+++ tests/test_cli.py
@@ -1,20 +1,18 @@
-from contextlib import ExitStack as does_not_raise  # noqa: N813
+import subprocess
+import textwrap
 
-import click
 import pytest
-from pytask_parallel.cli import _validate_n_workers
 
 
-@pytest.mark.unit
-@pytest.mark.parametrize(
-    "value, expectation",
-    [
-        (0, pytest.raises(click.UsageError)),
-        (1, does_not_raise()),
-        (2, does_not_raise()),
-        ("auto", does_not_raise()),
-    ],
-)
-def test_validate_n_workers(value, expectation):
-    with expectation:
-        _validate_n_workers(None, None, value)
+@pytest.mark.end_to_end
+def test_delay_via_cli(tmp_path):
+    source = """
+    import pytask
+
+    @pytask.mark.produces("out_1.txt")
+    def task_1(produces):
+        produces.write_text("1")
+    """
+    tmp_path.joinpath("task_dummy.py").write_text(textwrap.dedent(source))
+
+    subprocess.run(["pytask", tmp_path.as_posix(), "-n", "2", "--delay", "5"])